==> 1 - in-house_raw_data_preprocessing_new.Rmd <==

# Library:
```{r setup, include=FALSE}
library(ChAMP)
library(dplyr)
library(lumi)
library(minfi)
```

# 1. Load data
```{r}
#load raw idat file with ChAMP pipeline, turn off filteration for SNPs, XY chromesome and multihit probes
myLoad<-champ.load("/nfs/dcmb-lgarmire/liuwent/01-data", arraytype="EPIC", filterSNPs=FALSE, filterXY=FALSE, filterMultiHit = FALSE)
dim(myLoad$beta)#827573

#remove multi-hit probes using Zhou et al.,2017 reference 
load("/nfs/dcmb-lgarmire/xtyang/CordBlood/multi_hit_new.RData")
myLoad$beta = myLoad$beta[!rownames(myLoad$beta)%in%multi_hit_new$probeID, ]
dim(myLoad$beta) #803359 probes

#Simplify column names
colnames(myLoad$pd)[1]="Sample_Name"
colnames(myLoad$pd)[13]="GA"
colnames(myLoad$pd)[14]="Eth2"
```

## Organize ethnicities into bigger categories
```{r}
for (i in 1:nrow(myLoad$pd)){
  if(myLoad$pd$Eth2[i] %in% c("American Samoan", "Chamorro", "Micronesian","MicronesianChuukese", "MicronesianKosraean", "samoan", "Samoan", "Tongan", "Hawaiian", "Tahitian")){
    myLoad$pd$Eth2[i] = "PacificIslander"}
  
  else if(myLoad$pd$Eth2[i] %in% c("Asian", "Cambodian", "Chinese", "Filipino", "Japanese", "Korean", "Laotian", "Okinawan", "Vietnamese","American Indian")){
    myLoad$pd$Eth2[i] = "Asian"}
  
  else if(myLoad$pd$Eth2[i] %in% c("Caucasian", "English", "German", "Hungarian", "Irish", "Italian", "Portuguese", "Scottish","Spanish")){
    myLoad$pd$Eth2[i] = "Caucasian"}
  
  else{myLoad$pd$Eth2[i] = "Others"}
}
```

## Remove others group of Eth2
```{r}
myLoad$pd <- myLoad$pd[which(myLoad$pd$Eth2 != 'Others'), ]
dim(myLoad$pd)

myLoad$beta <- myLoad$beta[, which(colnames(myLoad$beta)%in%myLoad$pd$Sample_Name)]
dim(myLoad$beta)
```

## Missing BMI values imputation (use mean of each sample groups to impute):
```{r}
myLoad$pd$BMI[c(45,53)] = c(27.9, 27.9)
myLoad$pd$BMI[52] = 32.2
```


## 1.1 Normalization and remove batch effect

```{r}
#Density plot has irregular beta value, which needs to be removed with the following code
newpd<-myLoad$pd[!(myLoad$pd$Sample_Name)=="578",]
newbeta<-myLoad$beta[,!(colnames(myLoad$beta)==578)]

myLoad$pd<-newpd
myLoad$beta<-newbeta

pd<-myLoad$pd

#Normalization
myNorm0<-champ.norm(beta=myLoad$beta, arraytype="EPIC", cores = 8)

##Run QC on myNorm
champ.QC(beta=myNorm0, dendrogram = F, 
         resultsDir="/nfs/dcmb-lgarmire/xtyang/CordBlood/03-wenting_results/01-new_result_Garvin_ref/CHAMP_QCimages/")


#Save data:
pd<-myLoad$pd

save(myLoad, file = '/nfs/dcmb-lgarmire/xtyang/CordBlood/03-wenting_results/01-new_result_Garvin_ref/myLoad_new.RData')
save(myNorm0, file = '/nfs/dcmb-lgarmire/xtyang/CordBlood/03-wenting_results/01-new_result_Garvin_ref/myNorm_new.RData')
```
