==> 3.2 - SOV_Ching_data.R <==
# Library:
library(ggplot2)
library(EpiDISH)
library(parallel)
library(dplyr)
library(car)
library(lumi)

load("/nfs/dcmb-lgarmire/liuwent/13-Travers_data/Trav_pd_all_minfi.RData")
load("/nfs/dcmb-lgarmire//liuwent/13-Travers_data/myNorm2.RData")

m = beta2m(myNorm2)

#cell type deconvolution using Gervin's reference
load("/nfs/dcmb-lgarmire/xtyang/CordBlood/03-wenting_results/FlowSorted.CordBloodCombined.450k.compTable.rda")
ref = FlowSorted.CordBloodCombined.450k.compTable
out<-epidish(beta.m=beta, ref.m=ref, method="CP")
out$estF[out$estF<0]=0

#extract confounders
pd = Trav_pd_all_minfi
pd = pd%>%dplyr::select("Sample_Group",  "MomAge", "BMI", "EthnicMom1", "GAWeek", "BabySex")
pd = as.data.frame(cbind(pd, out$estF))
pd = pd%>%select(-nRBC)
#pd$Sample_Group = as.numeric(pd$Sample_Group)
pd$EthnicMom1 = as.factor(pd$EthnicMom1)
pd$BabySex = as.factor(pd$BabySex)

sovdat = data.frame(t(m))

Ftab <- data.frame(Sample_Group = numeric(), stringsAsFactors = FALSE)

for(i in 2:ncol(pd)){
  varname <- names(pd)[i]
  Ftab[varname] <- numeric()
}

calF <- function(probe = probecol){
  # Ensure sum contrasts for categorical variables
  newdata <- pd
  pdnames <- names(newdata)
  
  contrasts_list <- lapply(newdata, function(col) {
    if (is.factor(col)) return(contr.sum(length(levels(col))))
    return(NULL)
  })
  options(contrasts = c("contr.sum", "contr.poly"))
  
  
  newdata$beta <- probe
  formstr <- paste0(pdnames, collapse = ' + ')
  formstr <- paste0('beta ~ ', formstr)
  formstr <- as.formula(formstr)
  fit <- lm(formstr, data = newdata)
  aovfit <- Anova(fit, type = 3, singular.ok = TRUE)
  F <- aovfit$`F value`
  F <- F[2:(length(F)-1)]
  names(F) <- pdnames
  F <- as.data.frame(F, stringsAsFactors = FALSE)
  F <- as.data.frame(t(F))
  row.names(F) <- 1
  Ftab <- rbind(Ftab, F)
  return(Ftab)
}


Ftab <- mclapply(X = sovdat, FUN = calF, mc.cores = 16)
Ftab <- do.call(rbind, Ftab)
Fmean <- colMeans(Ftab)
Fmean <- Fmean[order(-Fmean)]
Fmean <- data.frame(Factor = names(Fmean), Fstat = as.vector(Fmean), stringsAsFactors = FALSE)
finalvars <- unique(c('Sample_Group', Fmean$Factor[Fmean$Fstat > 1]))

save(Ftab, Fmean, finalvars, file = "/nfs/dcmb-lgarmire/xtyang/CordBlood/03-wenting_results/02-Travis_data/sov_res_m.RData")

load("/nfs/dcmb-lgarmire/xtyang/CordBlood/03-wenting_results/02-Travis_data/sov_res_m.RData")

sovplot <- function(restab = MSSmean, clustername = 'Case', plottype = 'MSS',
                    textsize = 20){
  resmean <- restab
  samplegroupidx <- match('Sample_Group', resmean$Factor)
  resmean$Factor[samplegroupidx] <- paste0(clustername, '_Control')
  if(plottype == 'MSS'){
    ytitle <- 'Mean Square'
    resmean <- resmean[order(-resmean$MSSstat),]
    resmean$Factor <- factor(resmean$Factor, levels = resmean$Factor, ordered = TRUE)
    p <- ggplot(data = resmean, mapping = aes(x = Factor, y = MSSstat, fill = 'red'))
    print(
      p + geom_bar(stat = 'identity') +
        ggtitle('Source of Variance (Type 3 Anova)') +
        ylab(ytitle) +
        xlab('') +
        scale_fill_discrete(guide = FALSE) +
        theme_bw() +
        theme(axis.text.x = element_text(angle = 60, hjust = 1, size = textsize)) +
        theme(axis.text.y = element_text(size = textsize))
    )
  }else if(plottype == 'pval'){
    ytitle <- '-log2(p-val)'
    resmean <- resmean[order(-resmean$logpval),]
    resmean$Factor <- factor(resmean$Factor, levels = resmean$Factor, ordered = TRUE)
    
    p <- ggplot(data = resmean, mapping = aes(x = Factor, y = logpval, fill = 'red'))
    print(
      p + geom_bar(stat = 'identity') +
        ggtitle('Source of Variance (Type 3 Anova)') +
        ylab(ytitle) +
        xlab('') +
        scale_fill_discrete(guide = FALSE) +
        geom_hline(yintercept = -log2(0.05), color = 'red', size = 1) +
        theme_bw() +
        theme(axis.text.x = element_text(angle = 60, hjust = 1, size = textsize)) +
        theme(axis.text.y = element_text(size = textsize))
    )
  }else{
    ytitle <- 'F statistic'
    resmean <- resmean[order(-resmean$Fstat),]
    resmean$Factor <- factor(resmean$Factor, levels = resmean$Factor, ordered = TRUE)
    p <- ggplot(data = resmean, mapping = aes(x = Factor, y = Fstat, fill = 'red'))
    print(
      p + geom_bar(stat = 'identity') +
        ggtitle('Source of Variance (Type 3 Anova)') +
        ylab(ytitle) +
        xlab('') +
        scale_fill_discrete(guide = FALSE) +
        theme_bw() +
        theme(axis.text.x = element_text(angle = 60, hjust = 1, size = textsize)) +
        theme(axis.text.y = element_text(size = textsize))+
        geom_hline(yintercept=1, linetype="dashed", color = "red")
    )
  }
}

pdf("/nfs/dcmb-lgarmire/xtyang/CordBlood/03-wenting_results/02-Travis_data/SOV_Travis_Gervin_red.pdf")
sovplot(restab = Fmean, plottype = 'F', textsize = 14)
dev.off()

Fmean
#The variables that are over 1 are the confounding factors you will address.
